Factor analytic variation of EM algoritm

Implementation of the factor analytic variation of the EM algoritm as proposed by Dahl et al. (2013).

Usage

EMFA(
  y,
  k,
  size_param_x = NULL,
  cmHet = TRUE,
  dmHet = TRUE,
  tolerance = 1e-06,
  maxIter = 300L,
  size_param_cmStart = NULL,
  size_param_dmStart = NULL,
  mG = 1L,
  mE = 1L,
  maxDiag = 10000,
  stopIfDecreasing = TRUE,
  traits = ""
)

Arguments

y

An n x p matrix of observed phenotypes, on p traits or environments for n individuals. No missing values are allowed.

k

An n x n kinship matrix.

size_param_x

An n x c covariate matrix, c being the number of covariates and n being the number of genotypes. c has to be at least one (typically an intercept). No missing values are allowed. If not provided a vector of 1s is used.

cmHet

Should an extra diagonal part be added in the model for the precision matrix Cm?

dmHet

Should an extra diagonal part be added in the model for the precision matrix Dm?

tolerance

A numerical value. The iterating process stops if the difference in conditional log-likelihood between two consecutive iterations drops below tolerance.

maxIter

A numerical value for the maximum number of iterations.

size_param_cmStart

A p x p matrix containing starting values for the precision matrix Cm.

size_param_dmStart

A p x p matrix containing starting values for the precision matrix Dm.

mG

An integer. The order of the genetic part of the model.

mE

An integer. The order of the environmental part of the model.

maxDiag

A numical value. The maximal value of the diagonal elements in the precision matrices Cm and Dm (ignoring the low-rank part W W^t)

stopIfDecreasing

Should the iterating process stop if after 50 iterations the log-likelihood decreases between two consecutive iterations?

Value

A list containing the following components

Vg The genetic variance components matrix.

Ve The environmental variance components matrix.

References

Dahl et al. (2013). Network inference in matrix-variate Gaussian models with non-independent noise. arXiv preprint arXiv:1312.1622.

Zhou, X. and Stephens, M. (2014). Efficient multivariate linear mixed model algorithms for genome-wide association studies. Nature Methods, February 2014, Vol. 11, p. 407–409