Update W and P in EMFA algorithm

Update W and P used in the iteration process in the EMFA algorithm.

Usage

updateFA(
  y,
  wStart,
  pStart,
  wNew,
  pNew,
  m0,
  hetVar = FALSE,
  maxDiag = 10000,
  tolerance = 1e-04,
  maxIter = 100L,
  printProgress = FALSE
)

Arguments

hetVar

Should an extra diagonal part be added in the model for the precision matrix?

maxDiag

A numerical value for the maximum value of the diagonal of P.

tolerance

A numerical value. The iterating process stops if the sum of the difference for P and W between two steps gets lower than this value.

maxIter

A numerical value for the maximum number of iterations.

printProgress

Should progress be printed during iterations?

Y

An n x p matrix or data.frame.

WStart

A p x p matrix or data.frame containing starting values for W.

m

An integer. The order of the model.

PStart

A p x p matrix or data.frame containing starting values for P.